lazy load tailwind
